Okay, things are still pretty broken, but we've chugged along quite nicely. Today's beast was ditching my versions of data structures in favor of utilizing the LOFS structures. This has been incredibly painful and tedious, but so far the calculation code is at the very least compiling.
The remainder of readlofs.cpp functionality needs to first be
reimplemented, then make sure everything in run_cm1.cpp is behaving as
expected, and then finally finish up in integrate.cu for the rest of the
stuff that I bulk commented out. Needed to do that to help narrow down
the remaining usages of the old datastruct implementation.