Okay, so first off, some of the deallocators were commented out from previous attempts at bug fixing. So that's been reverted.
Second, spent a lot of time trying to figure out the best way to handle
this implementation of the staggered mesh macros. In LOFS-Read, the mesh
is only needed in the internal part of the domain of the scalar mesh, so
the mesh size and buffer size often mismatch. That's okay, because
things get put on the scalar grid and move on.

With the parcels though, we need our mesh to be the same size as our
buffers for things like calculating DX, so the solution in this case is
to request buffers the same size as the mesh itself. To make sure things
work with the staggered macros and we dont have memory access issues, NX
and NY are reduced by 2 and NZ by 1 before being sent to the parcel
routines. This should have the desired effect.