A number of changes to this commit. First, print statements were turned back on for debugging. I need to make this a namelist flag so that it's not hard coded. Will take care of that later.
Parcels exiting the domain is now fully and completely handled in a
somehwat "elegant" way, meaning no crashes, proper program exit, and
masked array values. If a parcel has exited the top of the domain (needs
to be extended to the sides too), its position and velocity are set to
the fill value. This fill value is the default NC_FILL_VALUE. If there
are parcels still in the domain, those are allowed to continue
integrating, and the fill value is just carried forward in time for the
parcels that did leave. If all parcels have exited the domain, but the
user requested more integration time than was needed, the program will
now gracefully exit.

The other major change is that I've edited the RK2 integration. The loop
structure was adopted from CM1, but in all honesty, made the code hard
to read. Now the individual RK2 steps execute linearly without control
flow, making the ability to check for the fill value much easier.